move scan to iceberg module
