+++ amoro-ams/src/main/java/org/apache/amoro/server/optimizing/UnKeyedTableCommit.java
@@ -29,12 +29,12 @@ import org.apache.amoro.hive.table.SupportHive;
 import org.apache.amoro.hive.utils.HivePartitionUtil;
 import org.apache.amoro.hive.utils.HiveTableUtil;
 import org.apache.amoro.hive.utils.TableTypeUtil;
+import org.apache.amoro.iceberg.Constants;
 import org.apache.amoro.op.SnapshotSummary;
 import org.apache.amoro.optimizing.OptimizingInputProperties;
 import org.apache.amoro.optimizing.RewriteFilesOutput;
 import org.apache.amoro.optimizing.RewriteStageTask;
 import org.apache.amoro.properties.HiveTableProperties;
-import org.apache.amoro.server.AmoroServiceConstants;
 import org.apache.amoro.server.utils.IcebergTableUtil;
 import org.apache.amoro.table.MixedTable;
 import org.apache.amoro.table.UnkeyedTable;
@@ -263,7 +263,7 @@ public class UnKeyedTableCommit {
     }
 
     RewriteFiles rewriteFiles = transaction.newRewrite();
-    if (targetSnapshotId != AmoroServiceConstants.INVALID_SNAPSHOT_ID) {
+    if (targetSnapshotId != Constants.INVALID_SNAPSHOT_ID) {
       long sequenceNumber = table.asUnkeyedTable().snapshot(targetSnapshotId).sequenceNumber();
       rewriteFiles.validateFromSnapshot(targetSnapshotId).dataSequenceNumber(sequenceNumber);
     }
@@ -350,11 +350,11 @@ public class UnKeyedTableCommit {
   private static Set<String> getCommittedDataFilesFromSnapshotId(
       UnkeyedTable table, Long snapshotId) {
     long currentSnapshotId = IcebergTableUtil.getSnapshotId(table, true);
-    if (currentSnapshotId == AmoroServiceConstants.INVALID_SNAPSHOT_ID) {
+    if (currentSnapshotId == Constants.INVALID_SNAPSHOT_ID) {
       return Collections.emptySet();
     }
 
-    if (snapshotId == AmoroServiceConstants.INVALID_SNAPSHOT_ID) {
+    if (snapshotId == Constants.INVALID_SNAPSHOT_ID) {
       snapshotId = null;
     }
