+++ amoro-ams/src/test/java/org/apache/amoro/server/optimizing/OptimizingTestHelpers.java
@@ -19,13 +19,13 @@
 package org.apache.amoro.server.optimizing;
 
 import org.apache.amoro.TableTestHelper;
-import org.apache.amoro.server.table.BasicTableSnapshot;
-import org.apache.amoro.server.table.KeyedTableSnapshot;
-import org.apache.amoro.server.table.TableSnapshot;
 import org.apache.amoro.server.utils.IcebergTableUtil;
 import org.apache.amoro.shade.guava32.com.google.common.collect.Lists;
+import org.apache.amoro.table.BasicTableSnapshot;
 import org.apache.amoro.table.KeyedTable;
+import org.apache.amoro.table.KeyedTableSnapshot;
 import org.apache.amoro.table.MixedTable;
+import org.apache.amoro.table.TableSnapshot;
 import org.apache.iceberg.AppendFiles;
 import org.apache.iceberg.DataFile;
 import org.apache.iceberg.DeleteFile;
