+++ amoro-ams/src/test/java/org/apache/amoro/server/optimizing/scan/TableFileScanHelperTestBase.java
@@ -23,6 +23,7 @@ import org.apache.amoro.catalog.CatalogTestHelper;
 import org.apache.amoro.catalog.TableTestBase;
 import org.apache.amoro.data.DataFileType;
 import org.apache.amoro.data.PrimaryKeyedFile;
+import org.apache.amoro.optimizing.scan.TableFileScanHelper;
 import org.apache.amoro.shade.guava32.com.google.common.collect.Lists;
 import org.apache.amoro.utils.ContentFiles;
 import org.apache.iceberg.ContentFile;
