+++ amoro-ams/src/test/java/org/apache/amoro/server/optimizing/scan/TestKeyedTableFileScanHelper.java
@@ -25,11 +25,13 @@ import org.apache.amoro.catalog.BasicCatalogTestHelper;
 import org.apache.amoro.catalog.CatalogTestHelper;
 import org.apache.amoro.data.ChangeAction;
 import org.apache.amoro.io.MixedDataTestHelpers;
+import org.apache.amoro.optimizing.scan.KeyedTableFileScanHelper;
+import org.apache.amoro.optimizing.scan.TableFileScanHelper;
 import org.apache.amoro.server.optimizing.OptimizingTestHelpers;
-import org.apache.amoro.server.table.KeyedTableSnapshot;
 import org.apache.amoro.server.utils.IcebergTableUtil;
 import org.apache.amoro.shade.guava32.com.google.common.collect.Lists;
 import org.apache.amoro.table.KeyedTable;
+import org.apache.amoro.table.KeyedTableSnapshot;
 import org.apache.amoro.table.TableProperties;
 import org.apache.amoro.utils.ExpressionUtil;
 import org.apache.iceberg.AppendFiles;
