+++ amoro-format-iceberg/src/main/java/org/apache/amoro/table/BasicTableSnapshot.java
@@ -16,35 +16,17 @@
  * limitations under the License.
  */
 
-package org.apache.amoro.server.optimizing.scan;
+package org.apache.amoro.table;
 
-import org.apache.iceberg.ContentFile;
-import org.apache.iceberg.DataFile;
-import org.apache.iceberg.expressions.Expression;
-import org.apache.iceberg.io.CloseableIterable;
+public class BasicTableSnapshot implements TableSnapshot {
+  private final long snapshotId;
 
-import java.util.List;
-
-public interface TableFileScanHelper {
-  class FileScanResult {
-    private final DataFile file;
-    private final List<ContentFile<?>> deleteFiles;
-
-    public FileScanResult(DataFile file, List<ContentFile<?>> deleteFiles) {
-      this.file = file;
-      this.deleteFiles = deleteFiles;
-    }
-
-    public DataFile file() {
-      return file;
-    }
-
-    public List<ContentFile<?>> deleteFiles() {
-      return deleteFiles;
-    }
+  public BasicTableSnapshot(long snapshotId) {
+    this.snapshotId = snapshotId;
   }
 
-  CloseableIterable<FileScanResult> scan();
-
-  TableFileScanHelper withPartitionFilter(Expression partitionFilter);
+  @Override
+  public long snapshotId() {
+    return snapshotId;
+  }
 }
